en proyectos freelance agregar servitec (https://servitecuruguay.com/), esta es una descripcion de lo que hace el sistema: It started as a system to record mechanic jobs in the field, aiming to maintain a history of completed work and manage payment for the services rendered.
The system includes an agenda feature that allows the administrator to schedule future visits, and it sends advance notifications to the administrator via push notifications, email, and WhatsApp messages.

To enhance web visibility, it has been decided to add a landing page that showcases information about the company and displays relevant photos of completed jobs. This addition will contribute to presenting a more professional and appealing image to potential clients visiting the website.